Add OSM raw data tile creation test rig
For each test data file this feeds the tile creator the corresponding
input file (which typically would contain exactly one source element)
and then generates a single tile from that. The output of that is then
compared with the expected reference.

Several of the tests expose errors in the current code, but the current
baseline is the status quo nevertheless to better see the effects of the
upcoming fixes.