Port to Clipper2
This is a 1:1 port, not making use of any of the new API yet.

This changes a number of test outputs due to minor numerical variations,
either within a tile or by moving nodes to a neighbouring tile close to
the boundary. Essentially the output is still the same though.