Fix broken compilation with Qt6.6 and MSVC2022 under Windows 10.
MSVC is not able to build the hash-table of class where the key is not fully defined previously. All function to play the hash-tables are now hosted in a dedicated container outside OsmPlacemarkData.

The previous code compiled fine with MSVC2019 (and also G++ or Clang). Sounds like a new restriction introduced in the new Microsoft compiler.

CCBUGS: 398582