Use standard Qt logging for debug messages
Marble originally used mDebug() in most cases for debugging, defined
in src/lib/marble/MarbleDebug.h and the corresponding .cpp source.
This may have been because, back in the old Qt3 or Qt4 days, there
was no control over Qt debug output so this class managed the debug
message enabling and output itself, based on the "--debug-info"
command line option.

Now Qt debug output can be controlled either by its standard settings
and environment variables, or by using the Plasma settings, so there
is no point in having mDebug() and its own settings.  Using qDebug()
also has the advantage that the function name is shown (if the Qt
message format requests it) without having to explicitly include
Q_FUNC_INFO in the debug message.

By setting compiler definitions in the CMakeLists.txt file in an
appropriate part of the Marble source tree, difefrent parts of the
library or application can use different logging categories.  This
allows more specific control over which messages are shown.

Simply define mDebug() as a synonym for qCDebug() with the configured
logging category.  It must be a macro so that the correct source
location is captured.