Make marker-color SimpleStyle GeoJSON parsing compliant
Despite not even doing anything with marker-color, Marble already
decided that it wants to do it in a way that is incompatible with
the SimpleStyle spec. Not only are valid SimpleStyle colors not
supported by Marble, but named colors, which SimpleStyle explicitly
says it does not support, are claimed to be supported as the
perceived benefit of this incompatibility.

In the spirit of making data portable between applications, fix
this by writing our own hex color parsing code that complies
with the SimpleStyle specification for colors.